+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to the "coderadio" extension will be documented in this file
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [v0.0.2] - 2022-10-13
+### Added
+- Volume controls added
+- Extension layout changed
+- Global state management and UI scope execution
+
 ## [v0.0.1] - 2020-10-30
 
 - Initial release
   - can stream coderadio (tested on gnu/linux and windows)
   - can change vlc_path in settings
   - cannot play stream if there is no internet connection
+